requirements.txt
# TÜBİTAK İP-2 AI Bot System - Python Dependencies
# Python 3.10+ required

# Core ML/RL
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0

# gRPC Communication
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0

# Monitoring & Logging
tensorboard>=2.14.0
wandb>=0.15.0  # Optional: cloud monitoring

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0  # Pretty console output
typer>=0.9.0  # CLI interface

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
